requirements.txt
# Core ML Dependencies
torch==2.1.2
torchvision==0.16.2
diffusers==0.23.1
transformers==4.36.2
huggingface_hub==0.19.4
tokenizers==0.14.1
numpy>=1.26.3
Pillow>=10.2.0
safetensors>=0.3.1
accelerate>=0.25.0

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.9
pydantic==2.6.1


# Video Processing
opencv-python-headless==4.8.1.78
imageio==2.33.0

# System Monitoring
psutil==5.9.8

# Logging and Monitoring
prometheus_client>=0.19.0
loguru==0.7.2

# GPU Optimization
bitsandbytes==0.41.3

runpod==1.3.0
python-dotenv>=1.0.0
ffmpeg-python==0.2.0
requests>=2.31.0
python-json-logger>=2.0.7